+++ backend/apps/oa_filing/services/case_import_service.py
@@ -13,7 +13,7 @@ from django.utils import timezone
 from apps.contracts.models import Contract
 from apps.core.models.enums import CaseType
 from apps.oa_filing.models import CaseImportPhase, CaseImportSession, CaseImportStatus
-from apps.oa_filing.services.oa_scripts.jtn_case_import import (
+from apps.oa_filing.services.oa_scripts.jtn.case_import import (
     JtnCaseImportScript,
     OACaseCustomerData,
     OACaseData,
