+++ backend/apps/oa_filing/services/oa_scripts/jtn/__init__.py
@@ -1,0 +1,1 @@
+"""金诚同达 OA 自动化脚本。"""
