+++ backend/apps/oa_filing/services/oa_scripts/jtn/case_import/__init__.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 # re-export 数据结构（保持外部 import 兼容）
-from ..jtn_case_import_models import (
+from ..models import (
     OAConflictData,
     OACaseCustomerData,
     OACaseData,
