+++ backend/apps/oa_filing/services/oa_scripts/jtn/case_import/detail_extractor.py
@@ -9,14 +9,14 @@ from typing import Any
 
 from playwright.sync_api import Page
 
-from ..jtn_case_import_models import (
+from ..models import (
     OACaseCustomerData,
     OACaseData,
     OACaseInfoData,
     OAConflictData,
     CaseSearchItem,
 )
-from .jtn_http_client import _BASE_URL, _DETAIL_URL_TEMPLATE, _MEDIUM_WAIT
+from .http_client import _BASE_URL, _DETAIL_URL_TEMPLATE, _MEDIUM_WAIT
 
 logger = logging.getLogger("apps.oa_filing.jtn_case_import")
