+++ backend/apps/oa_filing/services/oa_scripts/jtn/client_import/__init__.py
@@ -1,0 +1,5 @@
+"""金诚同达 OA 客户导入脚本。"""
+
+from .service import JtnClientImportScript, OACustomerData
+
+__all__ = ["JtnClientImportScript", "OACustomerData"]
