+++ backend/tests/ci/unit/test_jtn_case_import_login_cookie_reuse.py
@@ -5,9 +5,9 @@ from typing import Any
 import httpx
 import pytest
 
-import apps.oa_filing.services.oa_scripts.jtn_case_import.jtn_http_client as jtn_http_client
-import apps.oa_filing.services.oa_scripts.jtn_case_import.jtn_playwright_browser as jtn_playwright_browser
-from apps.oa_filing.services.oa_scripts.jtn_case_import import CaseListFormState, JtnCaseImportScript
+import apps.oa_filing.services.oa_scripts.jtn.case_import.http_client as jtn_http_client
+import apps.oa_filing.services.oa_scripts.jtn.case_import.playwright_browser as jtn_playwright_browser
+from apps.oa_filing.services.oa_scripts.jtn.case_import import CaseListFormState, JtnCaseImportScript
 
 
 def test_login_prefers_cached_http_cookies(monkeypatch: pytest.MonkeyPatch) -> None:
